web/electroT/solution/deploy/service/requirements.txt
click==8.0.3
Flask==2.0.2
Flask-Login==0.5.0
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
psycopg2-binary==2.9.2
Werkzeug==2.0.2
